WIP: Save

+++ Boticario.Backend/Modules/Products/Boticario.Backend.Modules.Products/Dto/ProductOperationDto.cs
@@ -4,7 +4,7 @@ namespace Boticario.Backend.Modules.Products.Dto
 {
     public class ProductOperationDto
     {
-        public long Sku { get; set; }
+        public int Sku { get; set; }
         public string Name { get; set; }
         public InventoryOperationDto Inventory { get; set; }
     }
